hyprland/uwsm: part 2 Switch from run-as-service (homebrew systemd-run wrapper) to `uwsm app`.
This also involved writing a new plugin for anyrun to correctly handle
applications by running them with `uwsm app`. It passes the path to the
desktop ID entry, as this also gives `uwsm app` context about Comment,
which is then used to populate the Description of the generated systemd
unit.